# Conflicts:
#	src/main/java/ru/lod/spbalert/model/GroupInfo.java

+++ src/main/java/ru/lod/spbalert/service/InfoAlertService.java
@@ -78,7 +78,7 @@ public class InfoAlertService {
             } else if (o2 == null) {
                 return -1;
             }
-            return Integer.compare(o1.getActual(), o2.getActual());
+            return Double.compare(o1.getActual(), o2.getActual());
         }
     }
 }
